pyproject.toml
[project]
name = "fastapi101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "boto3>=1.38.0",
    "fastapi[standard]>=0.135.1",
    "opentelemetry-exporter-otlp>=1.38.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.59b0",
    "opentelemetry-sdk>=1.38.0",
    "pika>=1.3.2",
    "prometheus-client>=0.22.0",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "pytest-cov>=7.1.0",
    "redis>=6.0.0",
    "sqlmodel>=0.0.37",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.13.1",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
pretty = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
check_untyped_defs = true
plugins = []
exclude = ["^alembic/versions/"]
